logic 파일 분리
